% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data}
\alias{sim_data}
\title{sim_data}
\usage{
sim_data(
  size_clusters = c(40, 40, 40),
  p_inf = 50,
  p_noise = 950,
  mu = 2,
  rho_unif_range = list(min = 0.1, max = 0.9),
  out_pct = 0,
  out_p = 50,
  out_unif_range = list(min = -6, max = 6),
  iseed = NULL
)
}
\arguments{
\item{size_clusters}{a vector containing the cluster sizes to be generated. The length of the vector corresponds to the number of clusters to be generated. The default is c(40, 40, 40).}

\item{p_inf}{number of informative variables describing the generated cluster structure. The default is 50.}

\item{p_noise}{number of noise variables which do not contribute to the cluster separation. The default is 950.}

\item{mu}{numeric number used for generating the mean of multivariate Gaussian distributions of clusters. The default is 2.}

\item{rho_unif_range}{interval used for generating covariance matrice of multivariate Gaussian distributions of clusters. The default is \code{list(min=0.1, max=0.9)}.}

\item{out_pct}{proportion of outliers. The default is 0.}

\item{out_p}{number of variables whose values are replaced by uniformly distributed numbers for outliers. The default is 50.}

\item{out_unif_range}{interval of uniform distribution for outliers. The default is \code{list(min=-6, max=6)}.}

\item{iseed}{seed of random number generator. The default is NULL.}
}
\value{
\item{x}{data matrix of the synthetic dataset.}
\item{y}{integer vector corresponding to the cluster membership before outlier contamination.}
\item{lb}{integer vector with cluster lables and outlier label (0).}
}
\description{
This function generates a synthetic dataset for the purpose of data clustering
in case of both outliers and noise variables are present.
}
\details{
Clusters are generated in the first \code{p_inf} informative variables with various characteristics following Gaussian models.
The clusters have different mean vectors and covariance matrix which is additionally randomly rotated.
Noise variables are generated following standard normal distributions.
Outliers are optionally generated by randomly choosing informative & noise variables and replacing them with uniformly distributed numbers.
}
\examples{
# Generate 3 clusters of equal sizes in the first 50 informative variables
# with 950 noise variables and 10\% of uniformly distributed outliers in 50 variables.

d <- sim_data(out_pct = 0.1, iseed = 1)

# cluster membership with uniformly distributed outliers in 0 cluster
table(d$lb)

}
